requirements.txt
asgiref==3.5.0
Django==4.0.1
django-ckeditor==6.2.0
django-etc==1.3.1
django-hitcount==1.3.5
django-js-asset==2.0.0
Pillow==9.0.1
sqlparse==0.4.2
tzdata==2022.1
